rename dynamic secure broadcast to bft_membership
