+++ src/lib.rs
@@ -9,6 +9,6 @@ pub mod orswot;
 pub mod actor;
 pub mod at2_impl;
 pub mod deterministic_secure_broadcast;
-pub mod dynamic_secure_broadcast;
+pub mod bft_membership;
 pub mod net;
 pub mod traits;
